coordinator_agent/pyproject.toml
[project]
name = "coordinator-agent"
version = "0.1.0"
description = "すべてのエージェント呼び出しの起点となるエージェント"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk==0.2.8",
    "google-adk==1.4.2",
    "nest-asyncio>=1.6.0",
    "python-dotenv>=1.1.0",
    "python-ulid>=3.0.0",
    "streamlit>=1.45.1",
]
